// types/user.ts
export interface IUserProfile {
    name: string;
    modes: string[];
    nominationCount: number;
    osuNominationCount: number;
    taikoNominationCount: number;
    catchNominationCount: number;
    maniaNominationCount: number;
    resetCount: number;
    resetsDoneByUserCount: number;
    qaCount: number;
    mapperCount: number;
    mapperPercentage: number;
    greenSevCount: number;
    greySevCount: number;
    redSevCount: number;
}